packages/pipeline/pyproject.toml
[tool.poetry]
name = "loocbooc-pipeline"
version = "0.1.0"
description = "Loocbooc 3D garment reconstruction pipeline — photo/pattern to physics-accurate 3D model"
authors = ["Loocbooc Engineering <engineering@loocbooc.com>"]
readme = "README.md"
packages = [{ include = "pipeline" }]

[tool.poetry.dependencies]
python = "^3.11"

# Core
numpy = "^1.26"
Pillow = "^10.0"
scipy = "^1.11"

# Image processing
opencv-python-headless = "^4.8"
scikit-image = "^0.22"

# 3D mesh processing
open3d = "^0.18"
trimesh = "^4.0"
numpy-stl = "^3.0"

# Pattern file parsing
ezdxf = "^1.2"

# Fabric background removal (optional, can skip for testing)
rembg = { version = "^2.0", optional = true }

# EXIF metadata
piexif = "^1.1"
exifread = "^3.0"

# Async job processing
celery = "^5.3"
redis = "^5.0"

# Export formats
pygltflib = "^1.16"

# Image similarity (for camera angle sorting)
imagehash = "^4.3"

# Logging and config
structlog = "^24.0"
pydantic = "^2.5"
pydantic-settings = "^2.1"

# HTTP for webhook callbacks
httpx = "^0.27"

[tool.poetry.extras]
background-removal = ["rembg"]

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
pytest-asyncio = "^0.23"
pytest-cov = "^4.1"
black = "^24.0"
ruff = "^0.3"
mypy = "^1.8"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]

[tool.black]
line-length = 100

[tool.ruff]
line-length = 100
select = ["E", "F", "I"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
